Stop disabling scrolling on about page
Not 100% sure this is the best fix, but currently the about page doesn't scroll vertically at all, and that's definitely not right. This can't be a problem with this theme generally (someone would have noticed before now) so something else weird is going on. Removing this line fixes it, but might break stuff elsewhere?