Adding a URL and a note that the place may be closed
It's hard to tell -- the official website of the church mentions the crypts,
but says "go look on Google Maps for information", and is about as helpful
as a leaflet from the Scarfolk Tourist Office. There's a couple of
newspaper articles, including one from March 2025, which say that it's
closed until further notice.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>