US Sm-Sw: Deletions
"Small Town USA" and "sms!!!" are way too vague.

Soka Uni is just a university.

Ste Genevieve is a whole town (I think I'd accept it if someone linked to
a decent description of what's interesting there, but as it is, I think we
should drop it).

Spinners Grill is permanently closed

Southern Star Brewing is just a brewery and taproom -- no tours or
"interesting" stuff.

Subpar Miniature Golf is just that. Again, not all that unusual.

Stade Saturdays may still be running, but there doesn't seem to be any
actual central information on the events. All I found was a bunch of local
new articles from various years.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>